Configuration changes to enable BT audio.
For some reason path probing fails with Bluetooth ALSA device, so
skip-probe needs to be yes. This introduces the bug that path
configurations are not created for the card, but this is not problem
currently since there is only one valid path anyway.

Also disable fmtx/fmrx for now, since they don't seem to work and
changing to either profile and restarting PulseAudio causes that the
ALSA device can't be opened with bluetooth profile.